moved styles to its own folder
