+++ client/app/screens/AdminLogin.tsx
@@ -2,14 +2,10 @@ import { useState, useEffect, useContext } from "react";
 import { StackNativeScreenProps } from "../interfaces/StackParamList";
 
 // Components
-import {
-  Text,
-  View,
-  TextInput,
-} from "react-native";
+import { Text, View, TextInput } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import MapButton from "../components/MapButton";
-import styles from "../components/Styles";
+import styles from "../styles/Styles";
 
 // Hooks
 import useFetch from "../hooks/useFetch";
@@ -42,11 +38,10 @@ export default function AdminLogin({ navigation }: Props) {
     };
     const url = "users/login";
     fetchData({ url, options });
-  };
+  }
 
   // unmount error solution: https://stackoverflow.com/questions/58038008/how-to-stop-memory-leak-in-useeffect-hook-react
- useEffect (() => {
-
+  useEffect(() => {
     let unmounted = false;
     if (!data) return;
 
@@ -62,7 +57,7 @@ export default function AdminLogin({ navigation }: Props) {
       unmounted = true;
     };
   }, [data]);
-  
+
   return (
     <SafeAreaView style={styles.container}>
       <Text style={styles.msg}>Maps can only be edited by Administers.</Text>
@@ -91,12 +86,12 @@ export default function AdminLogin({ navigation }: Props) {
           secureTextEntry
         />
       </View>
-      <View style = {styles.btnContainer}>
+      <View style={styles.btnContainer}>
         <MapButton
           label={"Sign in"}
           backgroundColor={"#f1b265"}
           handlePress={handleSignIn}
-         />
+        />
       </View>
 
       {loading && <Text>Loading...</Text>}
